+++ src/main/config/config.h
@@ -121,7 +121,7 @@ void targetConfiguration(struct master_s *config);
 
 #ifdef ASYNC_GYRO_PROCESSING
 uint32_t getPidUpdateRate(void);
-uint16_t getGyroUpdateRate(void);
+uint32_t getGyroUpdateRate(void);
 uint16_t getAccUpdateRate(void);
 uint16_t getAttiUpdateRate(void);
 uint8_t getAsyncMode(void);
